Downloads/efficepart/tools/converter/requirements.txt
Pillow>=10.0.0,<11.0.0
pdf2image>=1.16.3,<2.0.0
img2pdf>=0.4.4,<0.5.0
Flask>=3.0.0,<4.0.0
Werkzeug>=3.0.0,<4.0.0
svglib>=1.5.1
python-magic>=0.4.27
reportlab>=4.0.0